types: move all friendly int classes to constants.py
