+++ tpm2_pytss/__init__.py
@@ -8,4 +8,5 @@ except NotImplementedError:
 from .TCTILdr import *
 from .TCTI import TCTI
 from .types import *
+from .constants import *
 from .TSS2_Exception import TSS2_Exception
