+++ tpm2_pytss/tsskey.py
@@ -5,6 +5,7 @@ SPDX-License-Identifier: BSD-2
 import warnings
 from ._libtpm2_pytss import lib
 from .types import *
+from .constants import TPM2_ECC, TPM2_CAP, ESYS_TR
 from asn1crypto.core import ObjectIdentifier, Sequence, Boolean, OctetString, Integer
 from asn1crypto import pem
